remove stale storage-provisioner-gluster addon
The Gluster project has not has a release in a few years, maintenance
slowed down to the point it is almost standing still. Heketi as
component for deploying parts of the storage platform has been archived
in 2023.

Providing the storage-provisioner-gluster addon might give users the
wrong expectations. There is no guarantee Gluster is working with recent
minikube versions.