vfkit: Extract readPidfile() helper
This will be useful to implement stop and kill with signals.

Remove unneeded os.Stat() call. os.ReadFile() will return
os.ErrNotExists if the file does not exist, so we don't need extra
syscall. This is also fix theoretical race if the pidfile is created
after we check if the file exist.

The new helper is generic like checkPid() and can be reused later by all
drivers managing a pidfile.