vfkit: More robust Stop()
If setting vfkit state to "Stop" fails, we used to return an error.
Retrying the operation may never succeed.

Fix by falling back to terminating vfkit using SIGTERM. This should be
handled as graceful shutdown in vfkit if it is implemented correctly.

We can still fail if the pidfile is corrupted. Getting process from the
pid returns an error, but it is documented to "always succeed on unix".

In the case when we are sure the vfkit process does not exist, we remove
the pidfile immediately, avoiding leftover pidfile if the caller does
not call GetState() after Stop().

Sending a signal using a pidfile can work for any driver using a
pidfile, so it is implemented as a reusable function.