build: Use jammy based xcgo image for hyperkit (#20746)
* build: Fix docker run when selinux is enabled

When mouting volumes we need to use :Z suffix to allow access to the
mounted directory. This fixes docker run when running on Fedora an other
distros using selinux by default.

* build: Use jammy based xcgo image for hyperkit

Image was built from:
https://github.com/nirs/xcgo

Maybe we need to add this repo under the minikube organization or
include the code in minikube repo.

With this I could build hyperkit using:

    $ make hyperkit_in_docker
    $ file out/docker-machine-driver-hyperkit
    out/docker-machine-driver-hyperkit: Mach-O 64-bit x86_64 executable, flags:<NOUNDEFS|DYLDLINK|TWOLEVEL|PIE>