iso: Disable grub timeout
This speeds up machine boot by 5 seconds. The timeout may be helpful for
debugging boot issues but we don't have a way to access the serial
console for debugging currently.

Testing shows about 4-5 seconds speedup. In vfkit direct kernel boot is
slightly faster but simplifying the code is worth the tiny difference.

| driver | timeout | start time |
|--------|---------|------------|
| vfkit  |     5.0 |      24.01 |
| vfkit  |     0.0 |      19.90 |
|  qemu  |     5.0 |      29.46 |
|  qemu  |     0.0 |      24.28 |