iso: Improve build and clean up configuration (#20991)
* iso: Extract buildroot target

Beofre we can build the iso, we need to clone and configure buildroot.
This is required to run iso-menuconfig-{arch}.

* iso: Extract iso-prepare-% target

This target prepare for building an iso or running menuconfig. With this
change we can run the {iso,linux}-menuconfig-{x86_64,aarch64} targets
without buidling the entire iso.

* iso: Fix linux-menuconfig-% target

Previouly it worked only after building the entire iso. Now we make this
target without building the iso or running iso-menuconfig.

On the first run this downloads and builds lot for packages required to
run the linux-menuconfig target, but it is much shorter than buidling
the entire iso.

* iso: Simplify linux-menuconfig-%

Preveviously we copied the defconfig manauly to the beoard config file.
This can be done using the special linux-update-defconfig target.

With this change we don't need to keep KERNEL_VERSION in the Makefile,
making future upgrade easier.

* iso: Update buildroot configuration for aarch64

Run `make iso-menuconfig-aarch64` without making any changes updates the
buildroot config. It seems that there were manual changes in the config
which are overwritten when running iso-menuconfig. Removing the manual
changes to make it easier to edit the configuration with kconfig.

* iso: Update buildroot configuration for x86_64

Same as the aarch64 change to make it easier to configure using kconfig.

* iso: Update linux configuration for aarch64

Same as iso-menuconfig-aarch64, run `make linux-menuconfig-aarch64` and
exit without any change to update the config. This seems to change the
order, removing manual changes from the config. This will make it easier
to configure using kconfig in the future.

* iso: Update linux configuration for x86_64

Same as the aarch64 changes to make it easier to configure using kconfig
in the future.

* iso: Disable all platform for aarch64

We run on qemu virt machine or apple virtualization so we don't need
support for all kinds of embeded Arm boards. This reduces the arm64 iso
size from 410 MiB to 392 MiB.

* Updating ISO to v1.36.0-1751221996-20991

* Updating ISO to v1.36.0-1751315722-20991

---------

Co-authored-by: minikube-bot <minikube-bot@google.com>