build: Skip building hyperkit driver
The build is failing now with this error:

    GOOS=darwin CGO_ENABLED=1 go build \
	-ldflags="-X k8s.io/minikube/pkg/drivers/hyperkit.version=v1.36.0 ...   \
	-o out/docker-machine-driver-hyperkit k8s.io/minikube/cmd/drivers/hyperkit
    + failed=2

hyperkit needs the xcgo image which is unmaintained since Sep 2023. We
forked it to build hyperkit for minikube 1.36, but this image has now
many vulnerabilities.

This change only skip building hyperkit so we can recover it quickly if
someone can fix the build.