Add a default cluster DNS name config
It is very handy to be able to store the default cluster DNS name so
that I can fire up and teardown clusters easily without losing that
setting. You can think of this as a default value for the
`start --dns-domain <val>` flag, but only when starting a new cluster.

In my particular case, I would like to set my default cluster dna domain
to a subdomain under a domain that I control. E.g.
`cluster.wt.user.dev.example.com`. I want to do this to make it easier
to manage TLS certs for my dev cluster.

For context, the `start --dns-domain <val>` flag works like this:
  If you fire up a new cluster with that flag, you will get a new cluster
  with the domain name base. If you start an existing cluster, the
  config will be updated to the new domain name base and then started.

This config will work a little differently. The change implements a
config that will only affect newly started cluster. Here are some
examples to show that difference:

Newly started cluster example:
```
$ minikube config set DefaultClusterDNSDomain cluster2.local
$ minikube start
...
```

Stop and starting the cluster with `--dns-domain` flag:
```
$ minikube stop
...
$ minikub start --dns-domain cluster3.local
...
```

Stopping and starting a new cluster with a non-default domain name and
without the `--dns-domain` flag:
```
$ minikube stop
...
$ minikub start
...
```

The reason for this behavior is that I am not configuring the name of a
cluster. That is a cluster configuration option. I am setting the
default for that.

If I manually overrode the name of a cluster when started previously, I
don't want a default option to override my cluster configuration.

Before:
  User cannot cannot change the default DNS domain name for a cluster
  to something other than cluster.local.

After:
  The behavior of the minikube start flag `--dns-domain` does not
  change at all.

  User can set the default cluster domain name to whatever they want
  via the config option DefaultClusterDNSDomain. E.g.:
  `minikube config set DefaultClusterDNSDomain cluster.wt.users.dev.example.com`

  Any cluster started after that will have the DNS domain set the the
  configured value UNLESS they override it with the `--dns-domain` flag.