iso: Minimal kernel for arm64
Create default arm64 config and disable stuff that we cannot use in
a VM.

This chagne was generated by:

1. Create defualt arm64 config

       cd out/buildroot/output-aarch64/build/linux-6.6.95
       make ARCH=arm64 defconfig
       make ARCH=arm64 menuconfig
       (exit saving changes)

2. Disable features that we don't need in the minikube VM:

       - Platform suppport
	 - all platforms
       - Device drivers
         - Multimedia support
         - Sound support

3. Updated our linux defconfig

       cd out/buildroot/output-aarch64
       make linux-update-defconfig

4. Normalize the config

       make linux-menuconfig-aarch64
       (exit saving changes)

With this config qemu, vfkit, and krunkit boot with --no-kubernetes, and
graceful shutdown works in vfkit and krunkit (using --restful-uri).

We cannot start kubernetes yet since some features are not available in
the default architecture config.