hack: Replace go workspace with replace
Using go workspace breaks iso build, and may affect the builds in
unwanted ways, bringing dependencies from the hack module into the
minikube module.

Remove the go.work file and add a replace directive to the hack module
can access minikube packages from the local directory.

Some programs working directory was the soruce directory. With this
change all the programs and scripts in the hack module are run from the
hack directory. Paths updated to the programs can find files in the
minikube root directory.

Some updaters and benchmarks are broken in master and are still broken
with this change:

- make update-istio-operator-version
- make update-amd-gpu-device-plugin-version
- make cpu-benchmark-idle
- make cpu-benchmark-autopause
- make time-to-k8s-benchmark