iso: Fix `minikube stop` with vfkit and krunkit drivers (#21089)
* iso: Minimal kernel for arm64

Create default arm64 config and disable stuff that we cannot use in
a VM.

This chagne was generated by:

1. Create defualt arm64 config

       cd out/buildroot/output-aarch64/build/linux-6.6.95
       make ARCH=arm64 defconfig
       make ARCH=arm64 menuconfig
       (exit saving changes)

2. Disable features that we don't need in the minikube VM:

       - Platform suppport
	 - all platforms
       - Device drivers
         - Multimedia support
         - Sound support

3. Updated our linux defconfig

       cd out/buildroot/output-aarch64
       make linux-update-defconfig

4. Normalize the config

       make linux-menuconfig-aarch64
       (exit saving changes)

With this config qemu, vfkit, and krunkit boot with --no-kubernetes, and
graceful shutdown works in vfkit and krunkit (using --restful-uri).

We cannot start kubernetes yet since some features are not available in
the default architecture config.

* iso: Add configs removed by defualt config

This restores the configs removed by updating from the default
architecture config. These configs are required for kubernetes support.

After adding the removed configs, run `make linux-menuconfig-aarch64` to
normalize the config and remove multimedia and sound card support again.

* iso: Unbreak go packages build

Adding go.work seems to break podman build. The workspace is needed only
for running the update commands so let's disable it when building the
iso.

We may need much bigger change to ensur that the workspace is used only
when running the update go commands, or remove it. This change fixes
only the iso build.

* Updating ISO to v1.36.0-1752940814-21089

---------

Co-authored-by: minikube-bot <minikube-bot@google.com>