vfkit: Restore compatibility with x86_64
In #20833 we switch to --bootloader=efi, breaking vfkit on x86_64.
Switch to --bootloader=linux to restore support for older macs. This
also enable running vfkit basic tests in github runners.

This change does not revert #20833 since it contained other important
changes that we squashed during merge. We also use the new
--bootloader=linux instead of the legacy --kernel, --initrd, and
--cmdline flags.

Issues:
- On x86_64 using console=hvc0 breaks boot. Using console=ttyS0 works,
  but serial.log is always empty.